% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_key_hdr_boxplot.R
\name{draw_key_hdr_boxplot}
\alias{draw_key_hdr_boxplot}
\title{Draw legend key}
\usage{
draw_key_hdr_boxplot(data, params, size)
}
\arguments{
\item{data}{data}

\item{params}{parameters}

\item{size}{size of legend key}
}
\description{
draw legend key for HDR box plot
}
