% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghdr_data.R
\docType{data}
\name{faithful}
\alias{faithful}
\title{Example dataframe for gghdr}
\format{
This contains the 'Old Faithful Geyser' data used
in the examples in README and vignettes.

The variables are as follows:

 \itemize{
  \item eruptions. Eruption time in mins
  \item waiting. Waiting time to next eruption in mins
 }
}
\usage{
data(faithful)
}
\description{
Example dataframe for gghdr
}
\keyword{datasets}
