% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-hdr-boxplot.R
\name{geom_hdr_boxplot}
\alias{geom_hdr_boxplot}
\title{geom_hdr_boxplot}
\usage{
geom_hdr_boxplot(
  mapping = NULL,
  data = NULL,
  stat = "hdrcde",
  position = "dodge2",
  ...,
  varwidth = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  prob = c(0.5, 0.95, 0.99)
)
}
\arguments{
\item{mapping}{}

\item{prob}{}
}
\description{
geom_hdr_boxplot
}
