% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-hdr-rug.R
\name{geom_hdr_rug}
\alias{geom_hdr_rug}
\title{geom_hdr_rug}
\usage{
geom_hdr_rug(
  mapping = NULL,
  data = NULL,
  stat = "hdr",
  position = "identity",
  ...,
  varwidth = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  prob = c(0.5, 0.95, 0.99)
)
}
\arguments{
\item{mapping}{Default: NULL}

\item{data}{data}

\item{stat}{stat}

\item{position}{Default: "identity"}

\item{...}{...}

\item{varwidth}{width, Default: FALSE}

\item{na.rm}{Default: FALSE}

\item{show.legend}{Default: NA}

\item{inherit.aes}{Default: TRUE}

\item{prob}{PARAM_DESCRIPTION, Default: c(0.5, 0.95, 0.99)}
}
\value{
geom_hdr_rug
}
\description{
rug visualization for HDR plot
}
\examples{
library(ggplot2)

ggplot(faithful, aes(y = eruptions)) +
  geom_hdr_rug()
}
