% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-hdr-boxplot.R
\docType{data}
\name{GeomHdrBoxplot}
\alias{GeomHdrBoxplot}
\title{GeomHdrBoxplot}
\format{
An object of class \code{GeomHdrBoxplot} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 8.
}
\usage{
GeomHdrBoxplot
}
\description{
GeomHdrBoxplot
}
\keyword{internal}
