% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-prob.R
\name{guide_train.prob_guide}
\alias{guide_train.prob_guide}
\alias{guide_geom.guide_prob}
\title{Helper methods for guides}
\usage{
\method{guide_train}{prob_guide}(guide, scale, aesthetic)

\method{guide_geom}{guide_prob}(guide, layers, default_mapping)
}
\value{
A guide object
}
\description{
Helper methods for guides
}
\keyword{internal}
