% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-prob.R
\name{guide_prob}
\alias{guide_prob}
\title{Probability shade bar guide}
\usage{
guide_prob(title = waiver(), ...)
}
\arguments{
\item{title}{A character string or expression indicating a title of guide.
If \code{NULL}, the title is not shown. By default
(\code{\link[ggplot2:waiver]{waiver()}}), the name of the scale object or the name
specified in \code{\link[ggplot2:labs]{labs()}} is used for the title.}

\item{...}{Further arguments passed onto either
\code{\link[ggplot2]{guide_colourbar}} or \code{\link[ggplot2]{guide_legend}}}
}
\value{
A guide object
}
\description{
The guide shows the colour scales mapped to different probability coverage required for HDRs
}
