% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghdr_data.R
\docType{data}
\name{lane2}
\alias{lane2}
\title{Example dataframe for gghdr}
\format{
This contains the 'lane2' data from the 'hdrcde' package
and is used in the examples in README and vignettes.

The variables are as follows:

 \itemize{
  \item flow. a numeric vector giving the traffic flow in vehicles per
   lane per hour
  \item speed. a numeric vector giving he speed in miles per hour
 }
}
\usage{
data(lane2)
}
\description{
Example dataframe for gghdr
}
\keyword{datasets}
