% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-hdr.R
\name{stat_hdr}
\alias{stat_hdr}
\title{stat_hdrc (rug)}
\usage{
stat_hdr(
  mapping = NULL,
  data = NULL,
  geom = "hdr_rug",
  position = "dodge2",
  ...,
  coef = 1.5,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Default: NULL}

\item{data}{Default: NULL}

\item{geom}{PARAM_DESCRIPTION, Default: 'hdr_boxplot'}

\item{position}{PARAM_DESCRIPTION, Default: 'dodge2'}

\item{...}{...}

\item{coef}{Default: 1.5}

\item{na.rm}{Default: FALSE}

\item{show.legend}{Default: NA}

\item{inherit.aes}{Default: TRUE}
}
\description{
stat for hdr (rug)
}
\details{
Hyndman's (1996) density quantile algorithm is used for calculation.
}
\references{
Hyndman, R.J. (1996) Computing and graphing highest density
regions. \emph{American Statistician}, \bold{50}, 120-126.

Samworth, R.J. and Wand, M.P. (2010). Asymptotics and optimal bandwidth
selection for highest density region estimation.  \emph{The Annals of
Statistics}, \bold{38}, 1767-1792.

Wand, M.P., Marron, J S., Ruppert, D. (1991) Transformations in density
estimation. \emph{Journal of the American Statistical Association},
\bold{86}, 343-353.
}
