% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-hdrcde.R
\name{stat_hdrcde}
\alias{stat_hdrcde}
\title{stat_hdrcde}
\usage{
stat_hdrcde(
  mapping = NULL,
  data = NULL,
  geom = "hdr_boxplot",
  position = "dodge2",
  ...,
  coef = 1.5,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\description{
layer for hdrcde
}
