% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-hdr-rug.R
\name{geom_hdr_rug}
\alias{geom_hdr_rug}
\title{geom_hdr_rug}
\usage{
geom_hdr_rug(
  mapping = NULL,
  data = NULL,
  stat = "hdr",
  position = "dodge2",
  ...,
  varwidth = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  prob = c(0.5, 0.95, 0.99)
)
}
\arguments{
\item{varwidth}{width, Default: FALSE}

\item{prob}{PARAM_DESCRIPTION, Default: c(0.5, 0.95, 0.99)}
}
\value{
geom_hdr_rug
}
\description{
rug visualization for HDR plot
}
\examples{
ggplot(faithful, aes(y = eruptions)) +
  geom_hdr_rug()
}
