% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-hdr-rug.R
\docType{data}
\name{GeomHdrRug}
\alias{GeomHdrRug}
\title{GeomHdrRug}
\format{
An object of class \code{GeomHdrRug} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 7.
}
\usage{
GeomHdrRug
}
\description{
ggproto object for geom_hrdr_rug
}
\keyword{internal}
