% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghdr-package.R
\docType{package}
\name{gghdr-package}
\alias{gghdr}
\alias{gghdr-package}
\title{gghdr: Visualisation of Highest Density Regions}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Geom and Stat components to visualise highest density 
             regions from the 'hdrcde' package with 'ggplot2'.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{package}
