% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-hdr.R
\name{stat_hdr}
\alias{stat_hdr}
\title{Stat for hdr box and rug plot}
\usage{
stat_hdr(
  mapping = NULL,
  data = NULL,
  geom = "hdr_rug",
  position = "dodge2",
  ...,
  coef = 1.5,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Default: NULL}

\item{data}{Default: NULL}

\item{geom}{PARAM_DESCRIPTION, Default: 'hdr_boxplot'}

\item{position}{PARAM_DESCRIPTION, Default: 'dodge2'}

\item{...}{...}

\item{coef}{Default: 1.5}

\item{na.rm}{Default: FALSE}

\item{show.legend}{Default: NA}

\item{inherit.aes}{Default: TRUE}
}
\value{
A [ggplot2::Stat] representing the data transformations with required mappings for plotting HDRs using [geom_hdr_boxplot()] and [geom_hdr_rug()].
}
\description{
calculate components of hdr box and rug plot
}
