% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-hdr.R
\name{stat_hdr}
\alias{stat_hdr}
\title{stat_hdrcde}
\usage{
stat_hdr(
  mapping = NULL,
  data = NULL,
  geom = "hdr_rug",
  position = "dodge2",
  ...,
  coef = 1.5,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{geom}{PARAM_DESCRIPTION, Default: 'hdr_boxplot'}

\item{position}{PARAM_DESCRIPTION, Default: 'dodge2'}
}
\description{
stat for hdrcde
}
