% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-hdrcde.R
\name{stat_hdrcde}
\alias{stat_hdrcde}
\title{Stat for hdrcde}
\usage{
stat_hdrcde(
  mapping = NULL,
  data = NULL,
  geom = "hdr_boxplot",
  position = "dodge2",
  ...,
  coef = 1.5,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Default: NULL}

\item{data}{Default: NULL}

\item{geom}{PARAM_DESCRIPTION, Default: 'hdr_boxplot'}

\item{position}{PARAM_DESCRIPTION, Default: 'dodge2'}

\item{...}{...}

\item{coef}{Default: 1.5}

\item{na.rm}{Default: FALSE}

\item{show.legend}{Default: NA}

\item{inherit.aes}{Default: TRUE}
}
\value{
A [ggplot2::Stat] representing the data transformations with required mappings for plotting conditional HDRs using [geom_hdr_boxplot].
}
\description{
compute highest density regions continuously over some conditioned variable
}
